requirements.txt
streamlit
pandas
google-generativeai
langchain
langchain-google-genai
chromadb
python-magic
openpyxl
PyPDF2
langchain_community
tiktoken
streamlit_pdf_viewer